+++ R/test_skeleton.R
@@ -7,10 +7,10 @@
 #' @section Side effects:
 #'   Creates an R script file in the file system.
 #'
-#' @param fname Character scalar. Target R script file to open. If empty the file
+#' @param fname `character` scalar. Target R script file to open. If empty the file
 #'   currently open in the editor will be used.
-#' @param open Logical scalar. Should the test file be opened after it is created?
-#' @param sep Character scalar. Separator between \file{test} and \file{fname}
+#' @param open `logical` scalar. Should the test file be opened after it is created?
+#' @param sep `character` scalar. Separator between \file{test} and \file{fname}
 #'   when constructing the test file name. Should either be `"_"` or `"-"` for
 #'   compatibility with testthat.
 #'
@@ -22,7 +22,7 @@
 test_skeleton <- function(
   fname = NULL,
   open = TRUE,
-  sep = options('testthis.sep')
+  sep = options("testthis.sep")
 ){
   require_rstudio()
   ensure_testthat()
@@ -32,9 +32,9 @@ test_skeleton <- function(
   }
 
   if(file.exists(fname)){
-    message(sprintf('* File alread exists: %s', fname))
+    message(sprintf("* File alread exists: %s", fname))
   } else {
-    message(sprintf('* Creating `%s`', fname))
+    message(sprintf("* Creating `%s`", fname))
     title_name <- stringi::stri_sub(
       tools::file_path_sans_ext(basename(fname)), 6
     )