+++ R/testthis-package.R
@@ -6,7 +6,8 @@
 #' testthis also provides utility function to manage tests in subdirectories of
 #' the test/testthis directory.
 #'
-#' For details please refer to `vignette("testthis")`
+#' For details please refer to the
+#' [README](https://s-fleck.github.io/testthis/index.html)
 #'
 #' @name testthis
 #' @import assertthat
@@ -35,9 +36,34 @@
 #'  `testthis::use_integration_tests()`, `testthis::test_integration()`, etc...}
 #'}
 #'
+#'
+#' @section Testthis Tags:
+#'
+#' test_this tags are special comments that modify the behaviour of the
+#' functions supplied by this package. They are of the form `#* @tag <value>`.
+#' Please not that only some test_this tags really require a `<value>`.
+#'
+#' **Valid tags for script files in the /R/ dir (`pkg/R/*.R`)**
+#'
+#' * `@testfile <filename>`: manually specify associated test file. Should
+#' usually start with `test_`. This is used by `test_this()`, `lest_this()` and
+#' `open_testfile()`.
+#'
+#'
+#' **Valid tags or test files (`pkg/tests/testthat/test_*.R`)**
+#'
+#' * `@testing <functionname>`:  mark `functionname` as tested.
+#' Should usually go next the associated `test_that()` call. This is used by
+#' `test_coverage()`.
+#'
+#'
 #' @seealso [usethis::edit_r_profile()]
 #' @docType package
 "_PACKAGE"
+
+
+
+
 
 
 
@@ -58,6 +84,31 @@ if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
 
   toset <- !(names(op.testthis) %in% names(op))
   if(any(toset)) options(op.testthis[toset])
+
+  if (requireNamespace("crayon", quietly = TRUE) && crayon::has_color()){
+
+    style_error   <- crayon::make_style("#BB3333", colors = 256)
+    style_fatal   <- function(...) style_error(crayon::bold(...))
+    style_warning <- crayon::make_style("#EEBB50", colors = 256)
+    style_subtle  <- crayon::make_style(grDevices::grey(0.5), grey = TRUE)
+    style_accent  <- crayon::make_style("#ca2c92", colors = 256)
+    col_nchar     <- crayon::col_nchar
+
+  } else {
+    style_fatal   <- function(...) paste(...)
+    style_error   <- style_fatal
+    style_warning <- style_fatal
+    style_subtle  <- style_fatal
+    style_accent  <- style_fatal
+    col_nchar     <- function(...) nchar(...)
+  }
+
+  assign("style_fatal", style_fatal, envir = parent.env(environment()))
+  assign("style_error", style_error, envir = parent.env(environment()))
+  assign("style_warning", style_warning, envir = parent.env(environment()))
+  assign("style_subtle", style_subtle, envir = parent.env(environment()))
+  assign("style_accent", style_accent, envir = parent.env(environment()))
+  assign("col_nchar", col_nchar, envir = parent.env(environment()))
 
   invisible()
 }