+++ NEWS.md
@@ -1,9 +1,25 @@
+# testthis 1.1.0
+
+* fixed `read_testdata()` so that it works with R CMD CHECK (#11)
+* added `find_testdata()` which returns the path to the testdata directory
+* added `version` and `refhook` argument to `use_testdata()` (which is passed
+  on to `saveRDS()`). You can use `version = 2` to create testdata that is
+  compatible with R < 3.5.0.
+* added `test_index()` which displays an index / table of contents of all 
+  `test_that()` calls in a package's `tests/testthat` directory. Uses
+  source markers if called from RStudio.
+* renamed `get_test_coverage()` to just `test_coverage()`
+* added `.covrignore` support to `test_coverage()`
+* `read_testdata()` now works in R CMD Check thanks to JakeVestal (#11)
+
+
+
 # testthis 1.0.4
 
 * Import `parse_ns_file()` from pkgload instead of devtools
 * #* @testfile tag paths can now contain the script file extension `.R`
 * removed deprecated `lest_this()` function
-* various small bugfixes and improvements
+* various small bug fixes and improvements
 
 
 
@@ -11,7 +27,7 @@
 
 * Maintenance release to make vignette comply with new CRAN guidelines
 * test (`test_this()`, `test_subdir()`, etc..) functions now save all files in 
-  Rstudio and reload the package before running tests.
+  RStudio and reload the package before running tests.
 * `lest_this()` has been deprecated. `test_this()` now behaves like 
   `lest_this()`
   
@@ -22,7 +38,7 @@
 
 * added `test_all()` to run tests in all subdirectories
 * Use `usethis::proj_get()` instead of the `base_path` function argument for
-  compatbility with usethis.
+  compatibility with usethis.
 
 
 
@@ -40,8 +56,8 @@
 # testthis 1.0.0
 
 * Reworked an expanded infrastructure functions, such as `use_testdata()` and
-  `use_test_subdir()`. See readme for details.
-* `test_skeleton()` now honours the `#* @testfile` tag
+  `use_test_subdir()`. See README for details.
+* `test_skeleton()` now honors the `#* @testfile` tag
 * `open_test()` can now jump back and forth between source and test file, and
   does no longer automatically create test files.
 * testthis now depends on the new usethis package that replaces some of the